:construction: Fix missing app context in cli command
