+++ ckanext/dalrrd_emc_dcpr/jobs.py
@@ -1,15 +1,15 @@
 """Asynchronous jobs for EMC-DCPR"""
 
-import functools
-import contextlib
 import logging
 import typing
 
 from ckan import model
-from ckan.config.middleware import make_app
 from ckan.plugins import toolkit
 
-from . import email_notifications
+from . import (
+    email_notifications,
+    provide_request_context,
+)
 from .constants import (
     DatasetManagementActivityType,
     DcprManagementActivityType,
@@ -20,21 +20,8 @@ from .constants import (
 logger = logging.getLogger(__name__)
 
 
-def provide_request_context(func):
-    @functools.wraps(func)
-    def wrapped(*args, **kwargs):
-        app = make_app(toolkit.config)
-        with app._wsgi_app.test_request_context() as context:
-            result = func(context, *args, **kwargs)
-        return result
-
-    return wrapped
-
-
 def test_job(*args, **kwargs):
-    with _generate_request_context() as context:
-        logger.debug(f"inside test_job - {args=} {kwargs=}")
-        logger.debug(f"we now have a context: {context=}")
+    logger.debug(f"inside test_job - {args=} {kwargs=}")
 
 
 @provide_request_context
@@ -330,10 +317,3 @@ def _get_org_members(org_name: str) -> typing.List:
         if user.get("state") == "active":
             members.append(user_obj)
     return members
-
-
-@contextlib.contextmanager
-def _generate_request_context():
-    app = make_app(toolkit.config)
-    with app._wsgi_app.test_request_context() as context:
-        yield context
